+++ src/edwh/__about__.py
@@ -1,4 +1,7 @@
+"""
+This file keeps track of the current package version.
+"""
 # SPDX-FileCopyrightText: 2023-present Remco Boerma <remco.b@educationwarehouse.nl>
 #
 # SPDX-License-Identifier: MIT
-__version__ = "0.17.2"
+__version__ = "0.18.5"