+++ src/edwh/__init__.py
@@ -1,7 +1,17 @@
+"""
+This file exposes some functions so this tool can be used as a library.
+"""
+
 # SPDX-FileCopyrightText: 2023-present Remco Boerma <remco.b@educationwarehouse.nl>
 #
 # SPDX-License-Identifier: MIT
 
-from . import tasks
-from .tasks import check_env, get_env_value, set_env_value, TomlConfig, read_dotenv  # ... and more?
-from .helpers import *
+from . import tasks  # noqa: this exposes it as a library
+from .helpers import *  # noqa: this exposes it as a library
+from .tasks import (  # noqa: this exposes it as a library
+    TomlConfig,
+    check_env,
+    get_env_value,
+    read_dotenv,
+    set_env_value,
+)