+++ src/edwh/local_tasks/plugin.py
@@ -1,31 +1,46 @@
 """
 Extra namespace for plugin tasks such as plugin.add
 """
+import concurrent.futures
+import datetime as dt
+import concurrent.futures
+import datetime as dt
 import json
-import sys
+import re
 import typing
+from collections import OrderedDict
 from dataclasses import dataclass
-
-from invoke import task, Context
+from pprint import pprint
+import dateutil.parser
+import requests
+import yarl
+from invoke import Context, task
 from packaging.version import parse as parse_package_version
 from termcolor import colored
 
+import sys
 from ..helpers import VerboseLogger, NoopLogger
+
 from ..meta import (
-    _pip,
-    _get_available_plugins_from_pypi,
-    _parse_versions,
-    _gather_package_metadata_threaded,
-    _get_latest_version_from_pypi,
     Version,
+    _gather_package_metadata_threaded,
+    _get_available_plugins_from_pypi,
+    _get_latest_version_from_pypi,
+    _parse_versions,
+    _pip,
 )
 
 
-def _plugins(c: Context, pip_command=_pip()) -> list[str]:
+def list_installed_plugins(c: Context, pip_command=None) -> list[str]:
     """
     List installed edwh-plugins
     """
-    return c.run(f'{pip_command} freeze | grep edwh', hide=True, warn=True).stdout.strip().split("\n")
+    if not pip_command:
+        pip_command = _pip()
+
+    return c.run(f"{pip_command} freeze | grep edwh", hide=True, warn=True).stdout.strip().split("\n")
+from collections import OrderedDict
+    _get_available_plugins_from_pypi,
 
 
 @dataclass
@@ -48,7 +63,8 @@ class Plugin:
         self.requires_python = self.metadata["info"]["requires_python"]
 
     def __repr__(self):
-        return f"<EW Plugin: {self.clean_name}-{(self.installed_version if self.is_installed else self.latest_version) or '?'} {'installed' if self.is_installed else 'available'}>"
+        version = (self.installed_version if self.is_installed else self.latest_version) or "?"
+        return f"<EW Plugin: {self.clean_name}-{version} {'installed' if self.is_installed else 'available'}>"
 
     def __str__(self):
         return json.dumps(self.__dict__)
@@ -56,7 +72,12 @@ class Plugin:
     def print_details(self, verbose=False):
         if self.is_outdated:
             if verbose:
-                plugin_details = f"• {self.clean_name} ({self.installed_version} < {self.latest_version}) - {self.github_url} - Python {self.requires_python}"
+                plugin_details = (
+                    f"• {self.clean_name} "
+                    f"({self.installed_version} < {self.latest_version}) "
+                    f"- {self.github_url} "
+                    f"- Python {self.requires_python}"
+                )
             else:
                 plugin_details = (
                     f"• {self.clean_name} ({self.installed_version} < {self.latest_version}) - {self.github_url}"
@@ -65,7 +86,7 @@ class Plugin:
             print(
                 colored(
                     plugin_details,
-                    'yellow',
+                    "yellow",
                 )
             )
         elif self.is_installed:
@@ -79,7 +100,7 @@ class Plugin:
             print(
                 colored(
                     plugin_details,
-                    'green',
+                    "green",
                 )
             )
         else:
@@ -93,21 +114,19 @@ class Plugin:
             print(
                 colored(
                     plugin_details,
-                    'red',
+                    "red",
                 )
             )
 
 
-def get_installed_plugin_info(c: Context) -> list[Plugin]:
+def _gather_plugin_info(c: Context, plugin_names: list[str]) -> list[Plugin]:
     """
-    For all available plugins, get a Plugin instance with info
+    For all queried plugins (in `plugin_names`), get a Plugin instance with info.
     """
-    available_plugins = ["edwh"] + _get_available_plugins_from_pypi('edwh', 'plugins')
-    installed_plugins_raw = _plugins(c)
-    if not installed_plugins_raw or len(installed_plugins_raw) == 1 and installed_plugins_raw[0] == "":
-        raise ModuleNotFoundError("No 'edwh' packages found. That can't be right")
+    installed_plugins_raw = list_installed_plugins(c)
     installed_plugins = _parse_versions(installed_plugins_raw)
-    plugin_info = _gather_package_metadata_threaded(available_plugins)
+    plugin_names = [_require_affixes(_) for _ in plugin_names]
+    plugin_info = _gather_package_metadata_threaded(plugin_names)
 
     return [
         Plugin(
@@ -119,10 +138,20 @@ def get_installed_plugin_info(c: Context) -> list[Plugin]:
             else None,
             metadata=plugin_info.get(plugin),
         )
-        for plugin in available_plugins
+        for plugin in plugin_names
     ]
 
-    # return available_plugins, installed_plugins, plugin_info
+
+def gather_plugin_info(c: Context) -> list[Plugin]:
+    """
+    For all available plugins, get a Plugin instance with info
+    """
+    available_plugins = ["edwh", *_get_available_plugins_from_pypi("edwh", "plugins")]
+    installed_plugins_raw = list_installed_plugins(c)
+    if not installed_plugins_raw or len(installed_plugins_raw) == 1 and installed_plugins_raw[0] == "":
+        raise ModuleNotFoundError("No 'edwh' packages found. That can't be right")
+
+    return _gather_plugin_info(c, available_plugins)
 
 
 @task(name="list")
@@ -135,10 +164,9 @@ def list_plugins(c, verbose=False):
 
     :param verbose: should all info such as installed version always be shown?
     """
+    plugins = gather_plugin_info(c)
     logger = VerboseLogger if verbose else NoopLogger
     log = logger().log
-
-    plugins = get_installed_plugin_info(c)
 
     old_plugins = []
     not_all_installed = False
@@ -161,7 +189,8 @@ def list_plugins(c, verbose=False):
         print()
         print(
             colored(
-                f"Tip: not all plugins are installed. For example, try `edwh plugin.add {not_all_installed}` or `edwh plugin.add all`",
+                f"Tip: not all plugins are installed. "
+                f"For example, try `edwh plugin.add {not_all_installed}` or `edwh plugin.add all`",
                 "blue",
             )
         )
@@ -171,6 +200,10 @@ def _require_affixes(package: str, prefix="edwh-", suffix="-plugin"):
     """
     affix is 'an addition to the base form or stem of a word in order to modify its meaning or create a new word.'
     """
+    if package == "edwh":
+        # don't require affixes!
+        return package
+
     package = package.removeprefix(prefix).removesuffix(suffix)
     return f"{prefix}{package}{suffix}"
 
@@ -184,7 +217,7 @@ def add_all(c):
         c (Context): invoke ctx
     """
     pip = _pip()
-    plugins = _get_available_plugins_from_pypi('edwh', 'plugins')
+    plugins = _get_available_plugins_from_pypi("edwh", "plugins")
 
     plugins = " ".join(plugins)
     c.run(f"{pip} install {plugins}")
@@ -199,13 +232,13 @@ def remove_all(c):
         c (Context): invoke ctx
     """
     pip = _pip()
-    plugins = _get_available_plugins_from_pypi('edwh', 'plugins')
+    plugins = _get_available_plugins_from_pypi("edwh", "plugins")
 
     plugins = " ".join(plugins)
     c.run(f"{pip} uninstall --yes {plugins}")
 
 
-@task(aliases=('install',))
+@task(aliases=("install",))
 def add(c, plugin_name: str):
     """
     Install a new plugin
@@ -246,7 +279,7 @@ def update(c, plugin_name: str, version: str = None):
     c.run(f"{pip} install {plugin_name}=={version}")
 
 
-@task(aliases=('uninstall',))
+@task(aliases=("uninstall",))
 def remove(c, plugin_name: str):
     """
     Remove a plugin (or 'all')
@@ -261,4 +294,246 @@ def remove(c, plugin_name: str):
     # ensure the prefix and suffix exist, but not twice:
     plugin_name = _require_affixes(plugin_name)
 
-    c.run(f"{pip} uninstall {plugin_name}")
+    c.run(f"{pip} uninstall --yes {plugin_name}")
+
+
+GITHUB_RAW_URL = yarl.URL("https://raw.githubusercontent.com")
+
+
+def get_changelog(github_repo: str | yarl.URL):
+    if isinstance(github_repo, str):
+        github_repo = yarl.URL(github_repo)
+
+    github_repo = github_repo.path.removeprefix("/")  # e.g. educationwarehouse/edwh
+    changelog_url = GITHUB_RAW_URL / github_repo / "master/CHANGELOG.md"  # replace github.com with github raw
+
+    return requests.get(str(changelog_url)).text
+
+
+def get_changelogs_threaded(github_repos: dict[str, str]):
+    """
+    For any package in packages, gather its metadata from pypi
+    """
+    all_data: dict[str, str] = {}
+    with concurrent.futures.ThreadPoolExecutor() as executor:
+        repo_urls = list(github_repos.values())
+        for result, package in zip(executor.map(get_changelog, repo_urls), github_repos.keys()):
+            all_data[package] = result
+
+    return all_data
+
+
+def _filter_away_version(changelog_version: Version, _filter: str) -> bool:
+    """
+    If _filter is a Version and it's bigger than the selected row (via 'changelog_version'), the row should not be visible.
+    """
+    try:
+        filter_version = parse_package_version(_filter)
+        return changelog_version < filter_version
+    except Exception:
+        return False
+
+
+def _filter_away_date(date: dt.datetime, _filter: str) -> bool:
+    """
+    If _filter is a date and it's bigger than the selected row (via 'date'), the row should not be visible.
+    """
+    try:
+        return date < dateutil.parser.parse(_filter)
+    except Exception:
+        return False
+
+
+def _filter_away(version: Version, date: dt.datetime, _filter: str) -> bool:
+    """
+    If a filter is numeric, it's probably not a version or date (it could be parsed as one but we want other behavior).
+    If it is not numeric, try filtering away low version or dates.
+
+    Returns True if a row can be removed and False if it has to stay.
+    """
+    return (not _filter.isnumeric()) and (_filter_away_version(version, _filter) or _filter_away_date(date, _filter))
+
+
+def sort_versions(key_value: tuple[str, typing.Any]) -> Version:
+    """
+    Can be used as key=sort_versions in sort_and_filter_changelog
+    """
+    key, value = key_value
+
+    try:
+        version, date = key.split(" ")
+        return parse_package_version(version)
+    except Exception:
+        # something went wrong, return something so sorting doesn't crash:
+        return Version("0.0.0")
+
+
+def parse_changelog(markdown: str):
+    """
+    Parse our CHANGELOG.md to a dictionary of {version: {type: [list of changes]}}
+    where version is e.g. v0.18.5 (2023-06-06)
+    where type is e.g. Fix
+    """
+    # thanks ChatGPT
+    changelog = {}
+    current_version = None
+    current_category = None
+
+    lines = markdown.split("\n")
+    for line in lines:
+        if line.startswith("# Changelog"):
+            continue
+
+        version_match = re.match(r"^## (.+)", line)
+        if version_match:
+            version = version_match.group(1)
+            changelog[version] = {}
+            current_version = version
+            continue
+
+        category_match = re.match(r"^### (.+)", line)
+        if category_match:
+            category = category_match.group(1)
+            changelog[current_version][category] = []
+            current_category = category
+            continue
+
+        feature_match = re.match(r"^\* (.+)", line)
+        if feature_match:
+            feature = feature_match.group(1)
+            changelog[current_version][current_category].append(feature)
+
+    return changelog
+
+
+def to_date(key: str):
+    """
+    Convert a changelog key `v0.0.0 (2000-01-01)` to a dt.datetime
+    """
+    try:
+        _, date = key.split(" ", 1)
+        return dateutil.parser.parse(date.removeprefix("(").removesuffix(")"))
+    except:
+        return dateutil.parser.parse("2000-01-01")
+
+
+def to_version(key: str):
+    """
+    Convert a changelog key `v0.0.0 (2000-01-01)` to a Version(0.0.0)
+    """
+    try:
+        key, _ = key.split(" ", 1)
+        return parse_package_version(key)
+    except:
+        return Version("0.0.0")
+
+
+def sort_and_filter_changelog(changelog: dict, since: str = None):
+    """
+    Since can be:
+    - a number - amount of releases to show.
+    - a version number - show changes starting from that version.
+    - a date - show changes starting from that date.
+    - major, minor, patch - show changes starting from the latest release of that type.
+    """
+    filtered = {}
+
+    prev_major = prev_minor = prev_patch = 0
+
+    for idx, (k, v) in enumerate(changelog.items()):
+        version = to_version(k)
+        date = to_date(k)
+
+        # checks to stop:
+        if since == "major" and (version.major < prev_major):
+            break
+        elif since == "minor" and (version.minor < prev_minor or version.major < prev_major):
+            break
+        elif since == "patch" and (
+                version.micro < prev_patch or version.minor < prev_minor or version.major < prev_major
+        ):
+            break
+        elif since.isnumeric() and idx >= int(since):
+            break
+
+        # checks to skip:
+        elif _filter_away(version, date, since):
+            # skip!
+            continue
+
+        prev_major = version.major
+        prev_minor = version.minor
+        prev_patch = version.micro
+        # checks passed, add to output
+        filtered[k] = v
+
+    return OrderedDict(sorted(filtered.items(), reverse=True, key=sort_versions))
+
+
+COLORS = {
+    "fix": "yellow",
+    "feature": "green",
+    "documentation": "blue",
+}
+
+BOLD_RE = re.compile(r"((\*\*|__).+?(\*\*|__))")
+
+
+def colored_markdown(text: str):
+    """
+    Prettify a changelog line (makes ** bold).
+
+    todo: more than bold?
+    """
+    final = ""
+    for part in BOLD_RE.split(text):
+        if part.startswith("**") and part.endswith("**"):
+            part = colored(part.removeprefix("**").removesuffix("**"), attrs=["bold"])
+        final += part
+    return final
+
+
+def display_changelogs(changelogs: dict[str, OrderedDict]):
+    """
+    Final step of changelog(), uses the result of {package: sort_and_filter_changelog()}.
+    """
+    for package, history in changelogs.items():
+        print(colored(package, "red", attrs=["bold", "underline"]))
+        for version, changes in history.items():
+            print("-", version)
+            for change_type, change_descriptions in changes.items():
+                print("--", colored(change_type,
+                                    COLORS.get(change_type.lower(), "white")
+                                    ))
+                for change in change_descriptions:
+                    print("----", colored_markdown(change))
+
+
+@task(iterable=["plugin_names"])
+def changelog(ctx, plugin_names: list[str], since: str = "5", new: bool = False):
+    """
+    Show changelogs for edwh plugins.
+    by default, changelogs from all plugins are shown.
+    Since can be used to filter/limit changes. By default, the last 5 releases are shown.
+    Since can be a number (amount of changes), a date (show releases from that date),
+    a version (releases starting from that version) or
+    'major'/'minor'/'patch' to show releases since the latest version of that type.
+    if 'new' is True, show only changes for outdated packages.
+    """
+
+    if new:
+        outdated_plugins = []
+        info = _gather_plugin_info(ctx, outdated_plugins)
+    elif not plugin_names:
+        info = gather_plugin_info(ctx)
+    else:
+        info = _gather_plugin_info(ctx, plugin_names)
+
+    changelogs = get_changelogs_threaded(
+        {plugin.clean_name: plugin.metadata["info"]["project_urls"]["Source"] for plugin in info}
+    )
+
+    changelogs = {k: parse_changelog(v) for k, v in changelogs.items()}
+    changelogs = {k: sort_and_filter_changelog(v, since) for k, v in changelogs.items()}
+
+    display_changelogs(changelogs)